requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
onnx==1.20.1
onnxruntime-gpu==1.24.1
onnxslim==0.1.85
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.5
psutil==7.2.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch @ file:///D:/Code/torch/torch-2.9.0%2Bcu130-cp313-cp313-win_amd64.whl#sha256=7d83c2439d01aefc8ffea61cae2b8288cded5a90f60e034bc9830a7dc8029d84
torchvision @ file:///D:/Code/torch/torchvision-0.24.0%2Bcu130-cp313-cp313-win_amd64.whl#sha256=7b9b7b64761628a42f2273f9fa4975b0ed9c01c1a4f91123d743d2dcdad594f5
typing_extensions==4.15.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
urllib3==2.6.2
